requirements.txt
alembic==1.15.2
attrs==25.3.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.3
flasgger==0.9.7.1
Flask==3.1.0
Flask-JWT-Extended==4.7.1
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.0
mistune==3.1.3
openpyxl==3.1.5
packaging==25.0
pluggy==1.5.0
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
pytest==8.3.5
pytest-flask==1.3.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3
Flask-Cors==4.0.0
gunicorn==20.1.0
flask-wtf==1.2.2
wtforms==3.2.1
python-dotenv
email_validator
Faker
